회원관리

+++ zenith-project/app/Controllers/Api/ApiUserController.php
@@ -2,7 +2,8 @@
 namespace App\Controllers\Api;
 
 use CodeIgniter\API\ResponseTrait;
-use App\Models\Api\UserModel;
+use CodeIgniter\Shield\Models\UserModel;
+
 
 class ApiUserController extends \CodeIgniter\Controller 
 {
@@ -16,10 +17,14 @@ class ApiUserController extends \CodeIgniter\Controller
     }
     
     public function get($id = false) {
-        if ($id) {
-            $data = $this->userModel->find($id);
-        } else {
-            $data = $this->userModel->findAll();
+        if (strtolower($this->request->getMethod()) === 'get') {
+            if ($id) {
+                $data = $this->userModel->find($id);
+            } else {
+                $data = $this->userModel->findAll();
+            }
+        }else{
+            return $this->fail("잘못된 요청");
         }
 
         return $this->respond($data);
@@ -27,11 +32,24 @@ class ApiUserController extends \CodeIgniter\Controller
 
     protected function put($id = false) {
         $ret = false;
-        if ($id && !empty($this->data)) {
-            $ret = true;
-            $this->userModel->update($id, $this->data);
-        }
 
+        if (strtolower($this->request->getMethod()) === 'put') {
+            if ($id && !empty($this->data)) {
+
+                $this->validation = \Config\Services::validation();
+                $this->validation->setRules([
+                    'username' => 'required|is_unique[users.username]',
+                ]);
+                if($this->validation->run($this->data)){
+                    $ret = true;
+                    $this->userModel->update($id, $this->data);
+                }else{
+                    return $this->failValidationErrors("유효성 검사 에러");
+                }
+            }else{
+                return $this->fail("잘못된 요청");
+            }
+        }
         return $this->respond($ret);
     }
 
@@ -47,19 +65,22 @@ class ApiUserController extends \CodeIgniter\Controller
 
     protected function delete($id = false) {
         $ret = false;
-
-        if ($id) {
-            $ret = true;
-            $this->userModel->delete($id);
+        if (strtolower($this->request->getMethod()) === 'delete') {
+            if ($id) {
+                $ret = true;
+                $this->userModel->delete($id);
+            }
+        }else{
+            return $this->fail("잘못된 요청");
         }
-
+        
         return $this->respond($ret);
     }
 
     public function _remap(...$params) {
         $method = $this->request->getMethod();
         $params = [($params[0] !== 'get' ? $params[0] : false)];
-        $this->data = $this->request->getJSON();
+        $this->data = $this->request->getRawInput();
 
         if (method_exists($this, $method)) {
             return call_user_func_array([$this, $method], $params);
